refactor: move field focus functions to seperate file
